content/media/test/test_source_null.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=752087
-->
<head>
  <title>Test for Bug 752087</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=752087">Mozilla Bug 752087</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<video id="v">
<script>
var v = document.getElementById('v');
v.src = null; // crashes on NULL access if not handled

ok(true, "setting video.src to null didn't crash!");
</script>
</video>
</body>
</html>
